+++ test/unit/text/transformers/defaulttransformer.py
@@ -15,18 +15,26 @@
 # You should have received a copy of the GNU General Public License
 # along with Timeline.  If not, see <http://www.gnu.org/licenses/>.
 
+"""
+Unittest of class :doc:`DefaultTextTransformer
+<timelinelib_text_transformers_defaulttransformer>`
+"""
 
 from timelinelib.test.cases.unit import UnitTestCase
 from timelinelib.text.transformers.defaulttransformer import DefaultTextTransformer
 
 
 class describe_default_text_transformer(UnitTestCase):
+    """ """
 
     def test_simple_text_is_retruned_asis(self):
+        """ """
         self.assertEqual("", self.transformer.transform(""))
 
     def test_none_text_is_retruned_as_empty_string(self):
+        """ """
         self.assertEqual("", self.transformer.transform(None))
 
     def setUp(self):
+        """ """
         self.transformer = DefaultTextTransformer()